+++ prisma/migrations/20250825171332_remove_unique_constraint_from_path/migration.sql
@@ -1,0 +1,2 @@
+-- DropIndex
+DROP INDEX "Page_path_key";